+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-04-01
+
+### Added
+
+- `progress.column` option to show the cursor column in the secondary section
+
+### Fixed
+
+- Being able to use secondary section when following `mode`
+
 ## [1.3.0] - 2025-03-31
 
 ### Changed